@@ -1,33 +1,33 @@
 {
   "name": "yinyang-calendar.com",
   "version": "1.0.0",
-  "description": "今日が和暦で何月何日なのかが簡単にわかるカレンダー「Yin&Yang Calendar」",
-  "author": "Futa Ogawa",
   "private": true,
   "scripts": {
     "dev": "nuxt",
     "build": "nuxt build",
     "start": "nuxt start",
     "generate": "nuxt generate",
-    "lint": "eslint --ext .js,.vue --ignore-path .gitignore . --ignore-pattern 'docs/'",
-    "lintfix": "eslint --fix --ext .js,.vue --ignore-path .gitignore . --ignore-pattern 'docs/'",
-    "precommit": "npm run lint"
+    "lint:js": "eslint --ext .js,.vue --ignore-path .gitignore --ignore-path .eslintignore .",
+    "lintfix:js": "eslint --fix --ext .js,.vue --ignore-path .gitignore --ignore-path .eslintignore .",
+    "lint": "yarn lint:js",
+    "lintfix": "yarn lintfix:js"
   },
   "dependencies": {
-    "bootstrap": "^3.4.1",
-    "cross-env": "^5.2.0",
-    "nuxt": "^2.0.0"
+    "bootstrap": "3",
+    "core-js": "^3.6.5",
+    "nuxt": "^2.14.5"
   },
   "devDependencies": {
-    "babel-eslint": "^8.2.1",
-    "eslint": "^5.0.1",
-    "eslint-config-prettier": "^3.1.0",
-    "eslint-loader": "^2.0.0",
-    "eslint-plugin-prettier": "2.6.2",
-    "eslint-plugin-vue": "^4.0.0",
-    "node-sass": "^4.11.0",
-    "nodemon": "^1.11.0",
-    "prettier": "1.14.3",
-    "sass-loader": "^7.1.0"
+    "@nuxtjs/eslint-config": "^3.1.0",
+    "@nuxtjs/eslint-module": "^2.0.0",
+    "@nuxtjs/google-analytics": "^2.4.0",
+    "babel-eslint": "^10.1.0",
+    "eslint": "^7.8.1",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-nuxt": "^1.0.0",
+    "eslint-plugin-prettier": "^3.1.4",
+    "node-sass": "^4.14.1",
+    "prettier": "^2.1.1",
+    "sass-loader": "^10.0.2"
   }
 }
