@@ -4,7 +4,9 @@
       <div class="row">
         <div class="col-md-8 col-md-offset-2 footer">
           <div class="copyright">
-            <small>Copyright &copy; Yin&Yang Calendar All rights reserved.</small>
+            <small
+              >Copyright &copy; Yin&Yang Calendar All rights reserved.</small
+            >
           </div>
         </div>
       </div>
